ejemplo_django/requirements.txt
django
django-js-reverse
django-haystack
